Started on Validation class, added error popup
Started work on Validation class and I/O system for reading and writing
to files. Not functional as of now, but it probably will be in a few
days. An error window now pops up as a sort of exception handler. It
pops up when the login button is pressed right now as proof of it
existing, and the username/password variables in the login button
actionlistener are commented out.