Added FileWriter, Needs Testing
Added FileWriter and ability to create usernames. This needs to be tested, as something was wrong with the compiler on computer #32 or the wrong SDK was selected in IntelliJ. Also, a "Create Account" button needs to go in the loginPanel so the create new account functionality can actually be exercised. Once this is complete and an error frame is created for the username/password file not existing, begin work on the application.