Added new class for application
Added a class for the application, started on I/O utilities for timestamping and saving text from a JTextArea. Messed up brutally with Github Desktop and had to dig through a bunch of commits to recover this, so I'll probably have to delete all of the branches and recreate them after I merge this with master.