Added new class for application
Added a class for the application, started on I/O utilities for timestamping and saving text from a JTextArea. Messed up brutally with Github Desktop and had to dig through a bunch of commits to recover this. I'm deleting all of the branches and recreating them because of that, sorry guys (I do have a backup of everything on my hard drive, though).